django/requirements/main.txt
Django==4.0.5
psycopg2==2.9.3
tensorflow==2.9.1
spacy==3.3.1
tweepy==4.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
